--- app/__tests__/chathead/Step2Tutorial.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Step2Tutorial } from "../../src/client/chathead/Step2Tutorial";
- import ShallowRenderer from "react-test-renderer/shallow";
-
-
-describe("Tutorial", () => {
-    it("tests first content of tutorial", () => {
-
-    const renderer = new ShallowRenderer();
-    renderer.render(<Step2Tutorial />);
-    const result = renderer.getRenderOutput();
-
-    expect(result.type).toBe("section");
-    });
-});
